requirements.txt
py-linq
requests
bs4
sortedcontainers
pillow